Added JavaScript!! :D
Now it has a few databases and a spruced up Crafting tab to help me keep
track of things better. Next functionality is to figure out how to
condense the top tabs into a hamburger menu, and maybe to spruce up the
CSS of the crafting page in the future when item images come out.