kumo.html now uses two google spreadsheets
Gone are the days of manually manipulating the html. Now that there's a
google spreadsheet for both the individual user and the moderators, I am
one step closer to making this tracker incredibly easy to use for
non-web dev people. Also fixed the mod tools to no longer use canvasjs
due to the trial expiring.